backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.10
pinecone==5.3.0
pypdf==4.3.1
httpx==0.27.2
openai==1.55.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dateutil==2.9.0
